backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
bcrypt==4.1.2
catboost==1.2.3
cffi==1.16.0
click==8.1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
decorator==5.1.1
ecdsa==0.18.0
executing==2.0.1
fastapi==0.110.0
fonttools==4.49.0
graphviz==0.20.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.6
ipython==8.22.2
jedi==0.19.1
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pandas==2.2.1
parso==0.8.3
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
plotly==5.20.0
prompt-toolkit==3.0.43
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
tenacity==8.2.3
threadpoolctl==3.3.0
traitlets==5.14.2
typing_extensions==4.10.0
tzdata==2024.1
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
xgboost==2.0.3
